changed calculateDistance() return type from float to double

+++ app/src/androidTest/java/com/clockwork53/distancecalculator/distanceCalcTest.java
@@ -34,7 +34,7 @@ public class distanceCalcTest {
 
     @Test
     public void calculateDistance()  {
-        float actual = distanceCalcInstance.calculateDistance(a,b,c,d);
+        double actual = distanceCalcInstance.calculateDistance(a,b,c,d);
         Assert.assertEquals(e,actual,DELTA);
     }
